% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_extra.R
\name{celltypist_majority_vote}
\alias{celltypist_majority_vote}
\title{Celltypist Majority Vote}
\usage{
celltypist_majority_vote(
  data,
  classification_obs,
  groups_obs = NULL,
  graph = "RNA_snn",
  min_prop = 0,
  unassigned_label = "Unassigned"
)
}
\arguments{
\item{data}{A \code{Seurat} or \code{SingleCellExperiment} object containing the cell data.}

\item{classification_obs}{A string or list of strings specifying the column(s) in metadata where the cell type predictions (labels) are stored.}

\item{groups_obs}{A string specifying the column in metadata where the reference group labels are stored. If NULL, an over-clustering with the Leiden algorithm is performed based on the dataset size.}

\item{graph}{A string specifying the graph to use for clustering in \code{Seurat} objects. Default is "RNA_snn".}

\item{min_prop}{A numeric value specifying the minimum proportion of cells required to assign a majority vote label to a group. Default is 0.}

\item{unassigned_label}{A string specifying the label to assign to cell groups where no cell type reaches the minimum proportion. Default is 'Unassigned'.}
}
\value{
The modified \code{Seurat} or \code{SingleCellExperiment} object with new majority voting classification labels stored in metadata.
}
\description{
Assigns cell group labels based on the majority voting of cell type predictions within each group.
If no reference cell groups are provided, an over-clustering step is performed using the Leiden algorithm.
}
\details{
This function computes the majority voting for cell type labels within each group of cells.
If \code{groups_obs} is not provided, the function performs over-clustering using the Leiden algorithm with a resolution adjusted based on the dataset size.
The results of the majority voting are stored back in the metadata of the object, adding a column for each classification considered.
}
\examples{
\dontrun{
# Example usage with Seurat object
seurat_obj <- celltypist_majority_vote(
  seurat_obj,
  classification_obs = "predicted_labels",
  graph = "RNA_snn",
  min_prop = 0.5
)

# Example usage with SingleCellExperiment object
sce_obj <- celltypist_majority_vote(
  sce_obj,
  classification_obs = "predicted_labels",
  min_prop = 0.5
)
}
}
