% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA-pkg.R
\docType{package}
\name{CIA-pkg}
\alias{CIA}
\alias{CIA-package}
\alias{CIA-pkg}
\title{CIA}
\description{
Cluster Independent Annotation
}
\details{
This package provides functionality to ... TODO
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
