% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{compute_signature_scores}
\alias{compute_signature_scores}
\title{Compute Signature Scores}
\usage{
compute_signature_scores(
  data,
  geneset,
  seurat_assay = "RNA",
  matrix = "data",
  total_col_sums = NULL
)
}
\arguments{
\item{data}{Numeric matrix, data frame, SeuratObject, or SingleCellExperiment with genes in rows and cells in columns.}

\item{geneset}{Vector of gene names to compute the scores for.}

\item{seurat_assay}{Assay to use for SeuratObject objects (default "RNA").}

\item{matrix}{Slot to use for Seurat or SingleCellExperiment objects (default "data" for SO, 'logcounts' for SCE).}

\item{total_col_sums}{Optional precomputed column sums for normalization.}
}
\value{
Vector of signature scores for each column (cell) in the data.
}
\description{
Calculates signature scores for a given set of genes by comparing their expression levels
in the provided data. The score for each gene set is calculated as the product of the
count of genes expressed (presence) in a given cell and their relative expression levels (expression),
normalized by the total expression of the cell.
}
\examples{
## TODO
}
