% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{get_label}
\alias{get_label}
\title{Label extraction from the CIA score matrix}
\usage{
get_label(row, similarity_threshold, unassigned_label)
}
\arguments{
\item{row}{The vector of values, by row}

\item{similarity_threshold}{Numeric vector, to control the behavior of the
assignment of the labels.}

\item{unassigned_label}{The label to assign to the cells where no clear majority
signature is identified, default is "Unassigned" - handled via \code{CIA_classify}.}
}
\description{
Label extraction from the CIA score matrix
}
