% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_metrics.R
\name{group_composition}
\alias{group_composition}
\title{Group Composition Heatmap}
\usage{
group_composition(
  data,
  classification_obs,
  ref_obs,
  columns_order = NULL,
  color_map = "Greens"
)
}
\arguments{
\item{data}{Data frame or matrix containing the classification and reference data.}

\item{classification_obs}{Character string, the name of the column in \code{data} that contains the classification data.}

\item{ref_obs}{Character string, the name of the column in \code{data} that contains the reference group data.}

\item{columns_order}{Optional; a character vector specifying the order of columns in the heatmap.}

\item{color_map}{Optional; character string specifying the color palette to use, defaults to 'Greens'.
It should be one of the color palettes supported by \code{RColorBrewer}.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function generates a heatmap showing the percentage composition of each classification
within reference groups. It is used for visualizing how different classifications distribute
across predefined groups.
}
\examples{
## Assuming 'data' is a data frame with columns 'Group' and 'Category'
# group_composition(data, 'Category', 'Group')
# TODO

}
