% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_metrics.R
\name{grouped_classification_metrics}
\alias{grouped_classification_metrics}
\title{Compute Classification Metrics Per Reference Group}
\usage{
grouped_classification_metrics(
  cells_info,
  classification_col,
  ref_labels,
  unassigned_label = ""
)
}
\arguments{
\item{cells_info}{Data frame containing the observed data including the labels
assigned by the classification algorithm(s) and the true group labels.
Typically, this is the output of \code{colData()} for a SingleCellExperiment
object, and the \code{meta.data} slot for a Seurat object.}

\item{classification_col}{A strings specifying the column
name in \code{cells_info} where the labels assigned by the method of interest are
stored.}

\item{ref_labels}{Character string specifying the column in \code{cells_info} that
contains the true group labels (or the ones assigned by the reference method).}

\item{unassigned_label}{The label used in \code{classification_col} to denote
unclassified or unassigned samples.
If this parameter is specified and detected in the \code{cells_info} data, \%UN
will be calculated and included in the output metrics.}
}
\value{
A matrix with classification metrics (SE, SP, PR, ACC, F1) for each
reference group. If unassigned_label is specified and present in the `cells_info```
data, an additional '\%UN' column is included in the output.
}
\description{
Calculates classification metrics such as sensitivity (SE), specificity (SP),
precision (PR), accuracy (ACC), F1-score (F1), and the percentage of
'Unassigned' cells (\%UN),based on the
specified unassigned label, for each reference group. Cells labeled as specified by the value of
\code{unassigned_label} can be excluded from the metrics calculation.
Additionally, the percentage of unassigned cells for each classification
method is reported.
}
\details{
The classification metrics reported in the output object are;
\itemize{
\item the overall sensitivity (SE)
\item the specificity (SP)
\item the precision (PR)
\item the accuracy (ACC)
\item the F1-score (F1)
\item the percentage of unassigned cells (\%UN) for each classification method,
compared to the reference method.
}
}
