% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{load_signatures}
\alias{load_signatures}
\title{Load Gene Signatures}
\usage{
load_signatures(signatures_input, description_field_available = TRUE)
}
\arguments{
\item{signatures_input}{A character string representing the file path to the
TSV file containing the gene signatures.}

\item{description_field_available}{Logical value, to accommodate for the use
of "custom" gmt file formats, where the description field might not be
provided. Defaults to \code{TRUE}.}
}
\value{
A list where each element is a character vector of gene names, named
by the signature names.
}
\description{
This function loads gene signatures from a given input, which can either be a
file path to a TSV file or a named list.
Each signature in the file should be in a separate row with the first column
being the signature name and the remaining columns containing the genes.
}
\examples{
gmt_file <- system.file("extdata", "azimuth_human_motor_cortex.gmt",
                        package = "CIA")
signatures <- load_signatures(gmt_file)
signatures

gmt_url <-
  "https://data.wikipathways.org/20260110/gmt/wikipathways-20260110-gmt-Homo_sapiens.gmt"

sig_wikipathways <- load_signatures(gmt_url)
head(names(sig_wikipathways))
}
