% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_metrics.R
\name{plot_group_composition}
\alias{plot_group_composition}
\title{Plot Group Composition}
\usage{
plot_group_composition(
  df,
  ref_col,
  comp_col,
  plot_type = "percentage",
  palette = RColorBrewer::brewer.pal(8, "Set3"),
  show_legend = TRUE
)
}
\arguments{
\item{df}{DataFrame containing the data to be plotted.}

\item{ref_col}{Character. The name of the column representing the reference
grouping variable.}

\item{comp_col}{Character. The name of the column representing the grouping
to be compared.}

\item{plot_type}{Character. Indicates whether to plot 'percentage' or 'raw'
counts. Defaults to 'percentage'.}

\item{palette}{Character vector. Specifies the color palette to use. Defaults
to RColorBrewer::brewer.pal(8, "Dark2").}

\item{show_legend}{Logical. Whether to display the legend on the plot.
Defaults to TRUE.}
}
\value{
A horizontal stacked bar plot showing the composition of each
reference group.
}
\description{
This function plots the composition of each reference group as a horizontal
stacked bar plot. The composition can be shown either as raw counts or as
percentages.
}
\examples{
# TODO
# TODO: as an idea: we could flip the order of the levels to have it match the legend?

}
