% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{signature_score}
\alias{signature_score}
\title{Calculate Signature Scores}
\usage{
signature_score(
  data,
  signatures_input,
  return_score = FALSE,
  seurat_assay = "RNA",
  matrix = "data",
  score_mode = "raw",
  n_cpus = NULL
)
}
\arguments{
\item{data}{A numeric matrix, data frame, SeuratObject, or
SingleCellExperiment with genes in rows and cells in columns,
representing the expression level of each gene in each cell.}

\item{signatures_input}{A character string representing the file path to the
TSV file containing the gene signatures, or a list where each element is a
vector of gene names with names of the list elements representing signature
names.}

\item{return_score}{Boolean to return scores directly (default FALSE).}

\item{seurat_assay}{Assay for Seurat objects (default "RNA").}

\item{matrix}{Slot to use for Seurat or SingleCellExperiment objects
(default "data" for SO, "logcounts" for SCE).}

\item{score_mode}{Calculation mode for scores: "raw", "scaled", or
log-transformed ("log", "log2", "log10").}

\item{n_cpus}{Number of CPU cores for parallel computation (default uses a
quarter of available cores).}
}
\value{
Matrix of signature scores if return_score=TRUE or if data is a
matrix/data,frame. Otherwise, updates the input object's metadata with scores.
}
\description{
This function calculates the signature scores for each gene set in the
provided data. It loads the gene signatures from an input source and computes
the scores based on the expression levels of these genes in the data.
}
\examples{
## TODO

## TODO: rename this function? This actually does more than the previous one.
## Probably swapping the names is even a sensible choice?
}
