% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA-pkg.R
\docType{package}
\name{CIA-pkg}
\alias{CIA}
\alias{CIA-package}
\alias{CIA-pkg}
\title{CIA}
\description{
Cluster Independent Annotation
}
\details{
This package provides functionality to ... TODO
}
\author{
\strong{Maintainer}: a1 a2 \email{ab@c.d} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})

Authors:
\itemize{
  \item a1 a2 \email{ab@c.d} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})
  \item a1 a2 \email{ab@c.d} (\href{https://orcid.org/0000-0001-0002-0003}{ORCID})
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
}

}
\keyword{internal}
