% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA-pkg.R
\docType{package}
\name{CIA-pkg}
\alias{CIA}
\alias{CIA-package}
\alias{CIA-pkg}
\title{CIA}
\description{
Cluster Independent Annotation
}
\details{
\code{CIA} (Cluster Independent Annotation) is a cutting-edge computational tool
designed to accurately classify cells in scRNA-seq datasets using gene
signatures.
This tool operates without the need for a fully annotated reference dataset
or complex machine learning processes, providing a highly user-friendly and
practical solution for cell type annotation.

\code{CIA} summarizes the information of each signature expression into a single
score value for each cell.
By comparing these score values, \code{CIA} assigns labels to each cell based
on the top-scored signature.
\code{CIA} can filter scores by their distribution or significance, allowing
comparison of genesets with lengths spanning tens to thousands of genes.

\code{CIA} is implemented in both R and Python, making it compatible with all
major single-cell analysis frameworks like \code{SingleCellExperiment}, \code{Seurat},
and \code{Scanpy}.
This compatibility ensures a seamless integration into existing workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ingmbioinfo/CIA_R/}
  \item Report bugs at \url{https://github.com/ingmbioinfo/CIA_R/issues}
}

}
\author{
\strong{Maintainer}: Ivan Ferrari \email{ferrari@ingm.org} (\href{https://orcid.org/0000-0003-3746-4143}{ORCID})

Authors:
\itemize{
  \item Andrea Gobbini \email{gobbini@ingm.org} (\href{https://orcid.org/0000-0002-8846-4656}{ORCID})
  \item Mattia Battistella \email{battistella@ingm.org} (\href{https://orcid.org/0009-0006-9236-4398}{ORCID})
  \item Eugenia Galeota \email{galeota@ingm.org} (\href{https://orcid.org/0000-0002-4413-7389}{ORCID})
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
}

}
\keyword{internal}
