% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{CIA_classify}
\alias{CIA_classify}
\title{Classification Based on Signature Scores}
\usage{
CIA_classify(
  data,
  signatures_input,
  n_cpus = NULL,
  similarity_threshold = 0.1,
  seurat_assay = "RNA",
  seurat_layer = "data",
  sce_assay = "logcounts",
  score_mode = "scaled",
  column_name = "CIA_prediction",
  unassigned_label = "Unassigned"
)
}
\arguments{
\item{data}{A numeric matrix, data frame, \code{SeuratObject}, or
\code{SingleCellExperiment} with genes in rows and cells in columns,
representing the expression level of each gene in each cell.}

\item{signatures_input}{A character string representing the file path to the
TSV file containing the gene signatures, or a list where each element is a
vector of gene names with names of the list elements representing signature names.}

\item{n_cpus}{An optional integer indicating the number of CPU cores to use
for parallel computation. If NULL, the function will decide the number based on
available system resources.}

\item{similarity_threshold}{A numeric threshold used to decide if the highest
score is significantly higher than the second-highest to
assign a specific class to a cell. If the difference is less than or equal to
this threshold, the cell is labeled as unassigned.}

\item{seurat_assay}{Assay for Seurat objects (default "RNA").}

\item{seurat_layer}{Character string, indicating which layer to use of the
Seurat object. Defaults to "data".}

\item{sce_assay}{Character string, indicating which assay to use of the
\code{SingleCellExperiment} object. Defaults to "logcounts".}

\item{score_mode}{Character string, specifying the calculation mode to be
used for scores: "raw", "scaled", or log-transformed ("log", "log2", "log10").
In the scope of this function, this defaults to "scaled".}

\item{column_name}{The name of the column to be added to the metadata storing
the classification labels, default is "CIA_prediction".}

\item{unassigned_label}{The label to assign to the cells where no clear majority
signature is identified, default is "Unassigned".}
}
\value{
Modifies the input data object by adding the classification labels to
its metadata and returns the modified data object in the column specified by
\code{column_name}.
If input data is a matrix or data frame, this function returns a vector of
classification labels.
}
\description{
Performs classification of cells in \code{data} based on the computed signature
scores. Each sample/cell is classified to the signature with the highest score,
provided the difference between the top two scores exceeds a given threshold,
otherwise, it is marked as unassigned.
}
\examples{
## TODO example

}
\seealso{
\code{\link[=score_signature]{score_signature()}} \code{\link[=score_all_signatures]{score_all_signatures()}}
}
