% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_metrics.R
\name{compute_classification_metrics}
\alias{compute_classification_metrics}
\title{Compute Classification Metrics}
\usage{
compute_classification_metrics(
  data,
  classification_col,
  groups_cols,
  unassigned_label = ""
)
}
\arguments{
\item{data}{Data frame containing the observed data including the labels
assigned by the classification algorithm(s) and the true group labels.}

\item{classification_col}{A vector of character strings specifying the column
names in \code{data} where the labels assigned by the methods of interest are
stored.}

\item{groups_cols}{Character string specifying the column in \code{data} that
contains the true group labels (or the ones assigned by the reference method).}

\item{unassigned_label}{The label used in \code{classification_col} to denote
unclassified or unassigned samples.
If this parameter is specified and detected in the data, \%UN will be
calculated and included in the output metrics.}
}
\value{
A matrix with classification metrics (SE, SP, PR, ACC, F1) for each
classifier. If unassigned_label is specified and present in the data, an
additional '\%UN' column is included in the output.
}
\description{
Calculates classification metrics such as sensitivity (SE), specificity (SP),
precision (PR), accuracy (ACC), F1-score (F1), and the percentage of
'Unassigned' cells (\%UN) for given classification results, based on the
specified unassigned label. Cells labeled as specified by the value of
\code{unassigned_label} can be excluded from the metrics calculation.
Additionally, the percentage of unassigned cells for each classification
method is reported.
}
\details{
The classification metrics reported in the output object are;
\itemize{
\item the overall sensitivity (SE)
\item the specificity (SP)
\item the precision (PR)
\item the accuracy (ACC)
\item the F1-score (F1)
\item the percentage of unassigned cells (\%UN) for each classification method,
compared to the reference method.
}
}
\examples{
## TODO: probably have a pre-computed set of labels
## TODO: define-describe the process to get there in inst/scripts or so?

## # Assuming data is a data frame with true labels and classification results
## data <- read.csv('your_data_file.csv')
## classification_col <- c('classifier1', 'classifier2')
## groups_cols <- 'true_group'
## unassigned_label <- 'Unassigned' # Specify the label that denotes unassigned samples
## metrics_report <- compute_classification_metrics(data, classification_col, groups_cols, unassigned_label)
## print(metrics_report)
}
