% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{load_signatures}
\alias{load_signatures}
\title{Load Gene Signatures}
\usage{
load_signatures(signatures_input)
}
\arguments{
\item{signatures_input}{A character string representing the file path to the
TSV file containing the gene signatures, or a list where each element is a
vector of gene names with names of the list elements representing signature names.}
}
\value{
A list where each element is a character vector of gene names, named by the signature names.
}
\description{
This function loads gene signatures from a given input, which can either be a
file path to a TSV file or a named list.
Each signature in the file should be in a separate row with the first column
being the signature name and the remaining columns containing the genes.
}
\examples{
## TODO
}
