% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIA_core.R
\name{score_signature}
\alias{score_signature}
\title{Compute Signature Scores}
\usage{
score_signature(
  data,
  geneset,
  seurat_assay = "RNA",
  seurat_layer = "data",
  sce_assay = "logcounts",
  total_col_sums = NULL
)
}
\arguments{
\item{data}{Numeric matrix, data frame, SeuratObject, or SingleCellExperiment
with genes in rows and cells in columns.}

\item{geneset}{Vector of gene names to compute the scores for.}

\item{seurat_assay}{Assay to use for SeuratObject objects (default "RNA").}

\item{seurat_layer}{Character string, indicating which layer to use of the
Seurat object. Defaults to \code{data}.}

\item{sce_assay}{Character string, indicating which assay to use of the
SingleCellExperiment object. Defaults to \code{logcounts}.}

\item{total_col_sums}{Optional precomputed column sums for normalization.}
}
\value{
Vector of signature scores for each column (cell) in the data.
}
\description{
Calculates signature scores for a given set of genes by comparing their
expression levels in the provided data. The score for each gene set is
calculated as the product of the count of genes expressed (presence) in a
given cell and their relative expression levels (expression), normalized by
the total expression of the cell.
}
\examples{
## TODO example

}
